Move platform checking to service.platform
